Revert "Walker v1"
