--- CSCI_363_Final-main/WindowsFormsApp1/WindowsFormsApp1.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.33214.272
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WindowsFormsApp1", "WindowsFormsApp1\WindowsFormsApp1.csproj", "{CA791AFE-5954-450A-BFF7-987CAA3DBD95}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{CA791AFE-5954-450A-BFF7-987CAA3DBD95}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{CA791AFE-5954-450A-BFF7-987CAA3DBD95}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{CA791AFE-5954-450A-BFF7-987CAA3DBD95}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{CA791AFE-5954-450A-BFF7-987CAA3DBD95}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {969D3D5A-45D5-4623-8939-39C118DC9FC1}
-	EndGlobalSection
-EndGlobal
