+++ WindowsFormsApp1/WindowsFormsApp1/Form_Login.cs
@@ -2,87 +2,58 @@
 using System.Collections.Generic;
 using System.ComponentModel;
 using System.Data;
+using System.Diagnostics;
 using System.Drawing;
 using System.Linq;
 using System.Text;
+using System.Text.RegularExpressions;
 using System.Threading.Tasks;
 using System.Windows.Forms;
 
-namespace WindowsFormsApp1
-{
-    public partial class Form_Login : Form
-    {
-        public Form_Login()
-        {
+namespace WindowsFormsApp1 {
+    public partial class Form_Login : Form {
+        private bool isStudent, isProvider, isSuper;
+        public Form_Login() {
             InitializeComponent();
+            isStudent = isProvider = isSuper = false;
         }
 
-        private void label1_Click(object sender, EventArgs e)
-        {
+        private void button1_Click(object sender, EventArgs e) {
+            Regex student = new Regex("[A-Za-z0-9]+@student", RegexOptions.IgnoreCase);
+            Regex provider = new Regex("[A-Za-z0-9]+@provider", RegexOptions.IgnoreCase);
+            Regex super = new Regex("[A-Za-z0-9]+@super", RegexOptions.IgnoreCase);
 
-        }
 
-        private void label3_Click(object sender, EventArgs e)
-        {
 
-        }
-
-        private void Form_Login_Load(object sender, EventArgs e)
-        {
-
-        }
-
-        private void button1_Click(object sender, EventArgs e)
-        {
-            String errorMessage = "Username or Password is invalid";
-            if (textBox1.Text == "Student")
-            {
-                if (textBox2.Text == "1234")
-                {
-                    this.Hide();
-                    Form_Home home = new Form_Home();
-                    home.ShowDialog();
-                    this.Close();
-                }
-                else
-                {
-                    Login_Error_Message.Text = errorMessage;
-                }
-
+            if(student.IsMatch(textBox1.Text)) {
+                isStudent = true;
+            } else if(provider.IsMatch(textBox1.Text)) {
+                isProvider = true;
+            } else if(super.IsMatch(textBox1.Text)) {
+                isSuper = true;
+            } else {
+                textBox1.BackColor = Color.LightCoral;
+                return;
             }
-            else if (textBox1.Text == "Faculty")
-            {
-                if (textBox2.Text == "1234")
-                {
-                    this.Hide();
-                    Form_Home home = new Form_Home();
-                    home.ShowDialog();
-                
-                    this.Close();
-                }
-                else
-                {
-                    Login_Error_Message.Text = errorMessage;
-                }
 
+            Form_Home form = new Form_Home(isStudent, isProvider, isSuper);
+            form.Show();
+            this.Hide();
+        }
+
+        private void textBox1_TextChanged(object sender, EventArgs e) {
+            if(textBox1.BackColor == Color.LightCoral) {
+                textBox1.BackColor = SystemColors.Window;
             }
-            else if (textBox1.Text == "Admin")
-            {
-                if (textBox2.Text == "1234")
-                {
-                    this.Hide();
-                    Form_Home home = new Form_Home();
-                    home.ShowDialog();
-                    this.Close();
-                }
-                else
-                {
-                    Login_Error_Message.Text = errorMessage;
-                }
+
+            if(!string.IsNullOrEmpty(textBox2.Text)) {
+                button1.Enabled = true;
             }
-            else
-            {
-                Login_Error_Message.Text = errorMessage;
+        }
+
+        private void textBox2_TextChanged(object sender, EventArgs e) {
+            if(!string.IsNullOrEmpty(textBox1.Text)) {
+                button1.Enabled = true;
             }
         }
     }
